chore: move delegate and wallet component
