+++ src/modules/dpos/validator/components/delegatePerformance/styles.css
