+++ src/modules/dpos/validator/components/delegatePerformanceModal/styles.css
