+++ src/modules/dpos/validator/components/delegateProfile/delegateProfile.css
