+++ src/modules/dpos/validator/components/delegateProfile/index.js
