+++ src/modules/wallet/components/multiSignature/groupRow.js
