+++ src/modules/wallet/components/multiSignature/groupsTable.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import grid from 'flexboxgrid/dist/flexboxgrid.css';
-import Table from 'src/theme/table';
-import Box from 'src/theme/box';
-import BoxHeader from 'src/theme/box/header';
-import BoxContent from 'src/theme/box/content';
+import Table from '@theme/table';
+import Box from '@theme/box';
+import BoxHeader from '@theme/box/header';
+import BoxContent from '@theme/box/content';
 import GroupRow from './groupRow';
 import styles from './multiSignature.css';
