+++ src/modules/wallet/components/multiSignature/index.js
