+++ src/modules/wallet/components/multiSignature/mockData.js
