+++ src/modules/wallet/components/multiSignature/multiSignature.css
