+++ src/modules/wallet/components/multiSignature/multiSignature.js
