+++ src/modules/wallet/components/multiSignature/multiSignature.test.js
