+++ src/modules/wallet/components/multiSignature/transactionsTable.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Table from 'src/theme/table';
-import Box from 'src/theme/box';
-import BoxHeader from 'src/theme/box/header';
-import BoxContent from 'src/theme/box/content';
+import Table from '@theme/table';
+import Box from '@theme/box';
+import BoxHeader from '@theme/box/header';
+import BoxContent from '@theme/box/content';
 import TransactionRow from './transactionRow';
 import header from './transactionsTableHeader';
 import styles from './multiSignature.css';
