+++ src/modules/wallet/components/votes/index.js
