+++ src/modules/wallet/components/votes/votes.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   margin-bottom: 40px;
