+++ src/modules/wallet/components/votes/votes.js
@@ -2,11 +2,11 @@ import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import routes from 'src/routes/routes';
 import { isEmpty } from 'src/utils/helpers';
-import Box from 'src/theme/box';
-import BoxHeader from 'src/theme/box/header';
-import BoxContent from 'src/theme/box/content';
+import Box from '@theme/box';
+import BoxHeader from '@theme/box/header';
+import BoxContent from '@theme/box/content';
 import { Input } from 'src/theme';
-import Table from 'src/theme/table';
+import Table from '@theme/table';
 import styles from './votes.css';
 import VoteRow from './voteRow';
 import header from './votesTableHeader';
