+++ src/modules/wallet/components/votes/votes.test.js
