+++ src/modules/wallet/components/votes/votesTableHeader.js
